refactor menu - added separate menuReducer file
